feat: 결제 서비스에서 Kafka 기반 결제 처리 및 롤백 메시지 전송 로직 구현
- Kafka 토픽(SOLUTION_PAYMENT_TOPIC) 수신 후 결제 정보 저장 로직 구현
  - solutionSeq, amount, orderId, orderName 파싱 후 PaymentEntity 저장
- 예외 발생 시 solutionSeq 기반 Kafka 롤백 메시지(PAYMENT_SOLUTION_ROLLBACK_TOPIC) 전송
  - 메시지 파싱 로직 공통 처리 및 오류 상황 대응
