+++ src/main/java/startwithco/paymentservice/repository/PaymentEntityJpaRepository.java
@@ -1,8 +1,8 @@
-package startwithco.paymentservice.solutionPayment.repository;
+package startwithco.paymentservice.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import startwithco.paymentservice.solutionPayment.domain.PaymentEntity;
+import startwithco.paymentservice.domain.PaymentEntity;
 
 @Repository
 public interface PaymentEntityJpaRepository extends JpaRepository<PaymentEntity, Long> {
