+++ src/main/java/startwithco/paymentservice/repository/PaymentRepository.java
@@ -1,0 +1,7 @@
+package startwithco.paymentservice.repository;
+
+import startwithco.paymentservice.domain.PaymentEntity;
+
+public interface PaymentRepository {
+    PaymentEntity save(PaymentEntity payment);
+}
