--- src/main/java/startwithco/paymentservice/solutionPayment/repository/PaymentRepository.java
@@ -1,7 +1,0 @@
-package startwithco.paymentservice.solutionPayment.repository;
-
-import startwithco.paymentservice.solutionPayment.domain.PaymentEntity;
-
-public interface PaymentRepository {
-    PaymentEntity save(PaymentEntity payment);
-}
