+++ src/main/java/startwithco/paymentservice/topic/consumer/ConsumerTopic.java
@@ -1,4 +1,4 @@
-package startwithco.paymentservice.solutionPayment.kafka.topic.consumer;
+package startwithco.paymentservice.topic.consumer;
 
 public class ConsumerTopic {
     public static final String SOLUTION_PAYMENT_TOPIC = "solution-payment-topic";
